Fix tutorial code
- The `split` parameter to `as_tfdataset` is a kwarg.
- The attribute types are correct.

+++ docs/tutorials/quick_start/mnist_save.py
@@ -58,12 +58,12 @@ def main() -> None:
         Attribute(
             name="input",
             shape=(28, 28),
-            dtype="float16",
+            dtype="float32",
         ),
         Attribute(
             name="digit",
             shape=(),
-            dtype="int8",
+            dtype="uint8",
         ),
     ])
